+++ packages/browser-extension/src/modules/ui/utils/dom.ts
@@ -1,13 +1,18 @@
 export function selectElementContent(element: HTMLElement, endElement?: HTMLElement) {
-  let selection: Selection | null;
-  let range: Range;
+  const selection = window.getSelection();
+  const range = document.createRange();
 
-  selection = window.getSelection();
-  range = document.createRange();
+  const savedRange = selection?.rangeCount === 0 ? null : selection?.getRangeAt(0);
+
   range.selectNodeContents(element);
   if (endElement) {
     range.setEndAfter(endElement);
   }
+
   selection?.removeAllRanges();
   selection?.addRange(range);
+
+  return {
+    savedRange,
+  };
 }